refactor sparse adapter
